Begin central test mocking structure.
