+++ Tete.Tests/Setup/UserBase.cs
@@ -9,62 +9,8 @@ using Tete.Models.Users;
 
 namespace Tete.Tests.Setup
 {
-  public abstract class UserTestBase
+  public abstract class UserTestBase : TestBase
   {
-    protected Mock<Tete.Api.Contexts.MainContext> mockContext;
-
-    protected Guid existingUserId = Guid.NewGuid();
-
-    [SetUp]
-    public void Setup()
-    {
-      User existingUser = new User()
-      {
-        Id = existingUserId,
-        DisplayName = "test user",
-        UserName = "TestUser"
-      };
-
-      var userLanguage = new UserLanguage()
-      {
-        Language = new Language(),
-        UserId = existingUserId
-      };
-
-      var userProfile = new Profile(existingUserId);
-
-      var accessRoles = new List<AccessRole>() {
-        new AccessRole(existingUserId, "Admin") {
-          CreatedBy = existingUserId
-        }
-      };
-
-      IQueryable<User> users = new List<User> {
-        existingUser
-      }.AsQueryable();
-
-      IQueryable<UserLanguage> userLanguages = new List<UserLanguage> {
-        userLanguage
-      }.AsQueryable();
-
-      IQueryable<Profile> userProfiles = new List<Profile>() {
-        userProfile
-      }.AsQueryable();
-
-      IQueryable<AccessRole> userAccessRoles = accessRoles.AsQueryable();
-
-      var mockUsers = MockContext.MockDBSet<User>(users);
-      var mockUserLanguages = MockContext.MockDBSet<UserLanguage>(userLanguages);
-      var mockUserProfiles = MockContext.MockDBSet<Profile>(userProfiles);
-      var mockUserAccessRoles = MockContext.MockDBSet<AccessRole>(userAccessRoles);
-
-      mockContext = Tete.Tests.Setup.MockContext.GetDefaultContext();
-      mockContext.Setup(c => c.Users).Returns(mockUsers.Object);
-      mockContext.Setup(c => c.UserLanguages).Returns(mockUserLanguages.Object);
-      mockContext.Setup(c => c.UserProfiles).Returns(mockUserProfiles.Object);
-      mockContext.Setup(c => c.AccessRoles).Returns(mockUserAccessRoles.Object);
-
-    }
 
   }
 }
